🐛 move fetchParticipants()
