+++ app/lib/data.ts
@@ -4,23 +4,9 @@ import {
     Player
 } from '@/app/lib/definitions';
 import { sql } from '@vercel/postgres';
-import camelcaseKeys from "camelcase-keys";
 import decamelizeKeys from "decamelize-keys";
 
 
-export async function fetchParticipants() {
-    try {
-        const data = await sql<Player>`
-        SELECT * FROM players
-        WHERE is_player = TRUE
-        ORDER BY random();
-        `;
-        return camelcaseKeys(data.rows);
-    } catch (error) {
-        console.error('Database Error:', error);
-        throw new Error('Failed to fetch the players.');
-    }
-}
 
 export async function addPlayer(player: Player): Promise<void> {
     const dbPlayer = decamelizeKeys(player);
